requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
Werkzeug==2.3.7
pandas==2.0.3
numpy==1.24.3
yfinance==0.2.18
torch==2.0.1
transformers==4.33.2
accelerate==0.23.0
bitsandbytes==0.41.1
